+++ tests/shareable/test_fork.py
@@ -12,39 +12,6 @@ def fake_fcb_empty_case():
     return fcb
 
 
-@pytest.fixture(scope="session")
-def sample_forcing_config():
-    forcing_config = {
-        "basic": {
-            "dates": {
-                "start": "20200101",
-                "end": "20200109",
-                "format": "%Y%m%d",
-            },
-            "forcing": {
-                "product_name": "GLORYS",
-                "function_name": "get_glorys_data_script_for_cli",
-            },
-            "general": {"boundary_number_conversion": {"north": 1}},
-        },
-        "tides": {
-            "inputs": {
-                "tidal_constituents": ["M2", "K1"],
-                "boundaries": ["north"],
-                "tpxo_elevation_filepath": "ASd",
-                "tpxo_velocity_filepath": "ASd",
-                "case_specific_param": "asdsd",
-            }
-        },
-        "bgcic": {
-            "inputs": {
-                "marbl_ic_filepath": "qwreqwre",
-            }
-        },
-    }
-    return forcing_config
-
-
 def test_resolve_copy_plan_all_missing(fake_fcb_empty_case):
     """Test _resolve_copy_plan sets correct plan when all differences exist."""
     fcb = fake_fcb_empty_case
